Refactor 8 - planes
planes package has been created, within includes :
- FighterPlane
- bosses
  - BaseBoss
  - Boss
  - ISD
- enemies
  - EnemyPlane
  - TieFighter
- user
  - UserPlane
  - XWing

Boss and ISD now also utilise ProjectileFactory for projectile initialisation

deleted BossProjectile and ISDProjectile due to redundancy
