+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/levels/LevelSW.java
@@ -1,6 +1,6 @@
 package dev.ccr.dmscw2024.levels;
 
-import dev.ccr.dmscw2024.enemies.TieFighter;
+import dev.ccr.dmscw2024.planes.enemies.TieFighter;
 import dev.ccr.dmscw2024.fundamentals.LevelParent;
 
 public class LevelSW extends LevelParent {
