+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/planes/enemies/EnemyPlane.java
@@ -1,7 +1,7 @@
-package dev.ccr.dmscw2024.enemies;
+package dev.ccr.dmscw2024.planes.enemies;
 
 import dev.ccr.dmscw2024.fundamentals.ActiveActorDestructible;
-import dev.ccr.dmscw2024.fundamentals.FighterPlane;
+import dev.ccr.dmscw2024.planes.FighterPlane;
 import dev.ccr.dmscw2024.projectile.ProjectileFactory;
 
 public class EnemyPlane extends FighterPlane {
@@ -28,9 +28,7 @@ public class EnemyPlane extends FighterPlane {
 		if (Math.random() < FIRE_RATE) {
 			double projectileXPosition = getProjectileXPosition(PROJECTILE_X_POSITION_OFFSET);
 			double projectileYPosition = getProjectileYPosition(PROJECTILE_Y_POSITION_OFFSET);
-//			return new EnemyProjectile(projectileXPosition, projectileYPosition);
 			return ProjectileFactory.createProjectile("Enemy", projectileXPosition, projectileYPosition);
-
 		}
 		return null;
 	}
